Algebra Classes Restructured
The algebra class needs to be structured in such a way that it assumes there is a parent class converting the wstring input into a string input. This makes it easier for future classes so that they may work under the assumption that a class will be given a string. The parent class will also handle the conversion from wstring to string, so some functions were ommited.